Added: refresh token now automatically requested if access_token is expired

+++ AngularJSAuthentication/AngularJSAuthentication.Web/app/services/authService.js
@@ -2,6 +2,7 @@
 app.factory('authService', ['$http', '$q', 'localStorageService', function ($http, $q, localStorageService) {
 
     var serviceBase = 'http://localhost:52400/';
+    var $http;
     var authServiceFactory = {};
 
     var _authentication = {
@@ -14,6 +15,7 @@ app.factory('authService', ['$http', '$q', 'localStorageService', function ($htt
 
         _logOut();
 
+        $http = $http || $injector.get('$http');
         return $http.post(serviceBase + 'api/account/register', registration).then(function (response) {
             return response;
         });
@@ -30,6 +32,7 @@ app.factory('authService', ['$http', '$q', 'localStorageService', function ($htt
 
         var deferred = $q.defer();
 
+        $http = $http || $injector.get('$http');
         $http.post(serviceBase + 'token', data, { headers: { 'Content-Type': 'application/x-www-form-urlencoded' } }).success(function (response) {
 
             if (loginData.useRefreshTokens) {
@@ -72,33 +75,33 @@ app.factory('authService', ['$http', '$q', 'localStorageService', function ($htt
             _authentication.useRefreshTokens = authData.useRefreshTokens;
         }
 
-    }
-
+    };
 
     var _refreshToken = function () {
+
         var deferred = $q.defer();
 
         var authData = localStorageService.get('authorizationData');
 
-        if (authData) {
+        if (authData && authData.useRefreshTokens) {
 
-            if (authData.useRefreshTokens) {
+            var data = "grant_type=refresh_token&refresh_token=" + authData.refreshToken + "&client_id=" + "ngAuthApp";
 
-                var data = "grant_type=refresh_token&refresh_token=" + authData.refreshToken + "&client_id=" + "ngAuthApp";
+            localStorageService.remove('authorizationData');
 
-                localStorageService.remove('authorizationData');
+            $http = $http || $injector.get('$http');
+            $http.post(serviceBase + 'token', data, { headers: { 'Content-Type': 'application/x-www-form-urlencoded' } }).success(function (response) {
 
-                $http.post(serviceBase + 'token', data, { headers: { 'Content-Type': 'application/x-www-form-urlencoded' } }).success(function (response) {
+                localStorageService.set('authorizationData', { token: response.access_token, userName: response.userName, refreshToken: response.refresh_token, useRefreshTokens: true });
 
-                    localStorageService.set('authorizationData', { token: response.access_token, userName: response.userName, refreshToken: response.refresh_token, useRefreshTokens: true });
+                deferred.resolve(response);
 
-                    deferred.resolve(response);
-
-                }).error(function (err, status) {
-                    _logOut();
-                    deferred.reject(err);
-                });
-            }
+            }).error(function (err, status) {
+                _logOut();
+                deferred.reject(err);
+            });
+        } else {
+            deferred.reject();
         }
 
         return deferred.promise;
